+++ tests/core/responseTest.php
@@ -14,15 +14,17 @@ class ResponseTest extends TestCase
     public function setUp(): void
     {
         $this->res = new Response();
-        $this->res->write('This will not appear')->write('Hello')->append(' world');
         $this->res->setCode(301);
     }
 
 
     public function testInit()
     {
+        $this->res->write('This will not appear')->write('Hello')->append(' world');
         $this->assertEquals('Hello world', $this->res->get());
-        $this->assertEquals('Hello world', $this->res->send(true));
+        $this->res->writeJson([ 'msg' => 'Hello world' ]);
+        $this->assertEquals('{"msg":"Hello world"}', $this->res->get());
+        $this->assertEquals('{"msg":"Hello world"}', $this->res->send(true));
         $this->assertEquals(301, $this->res->getCode());
         $this->assertEquals(301, http_response_code());
         $this->assertInstanceOf(Response::class, $this->res->setHeader('Content-Type', 'plain/text'));